@@ -1,28 +1,28 @@
 {
-  "_from": "@schematics/angular@7.1.4",
-  "_id": "@schematics/angular@7.1.4",
+  "_from": "@schematics/angular@7.3.8",
+  "_id": "@schematics/angular@7.3.8",
   "_inBundle": false,
-  "_integrity": "sha512-4QVSmvQtOELek+FDq+k2ROeH9YrRrPJ6jWK179+qOruKSd4uTgEti/jy+fS0rfr52kDSGdDhz7XTh/QvQB89fg==",
+  "_integrity": "sha512-7o90bnIxXNpJhWPDY/zCedcG6KMIihz7a4UQe6UdlhEX21MNZLYFiDiR5Vmsx39wjm2EfPh3JTuBIHGmMCXkQQ==",
   "_location": "/@schematics/angular",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "@schematics/angular@7.1.4",
+    "raw": "@schematics/angular@7.3.8",
     "name": "@schematics/angular",
     "escapedName": "@schematics%2fangular",
     "scope": "@schematics",
-    "rawSpec": "7.1.4",
+    "rawSpec": "7.3.8",
     "saveSpec": null,
-    "fetchSpec": "7.1.4"
+    "fetchSpec": "7.3.8"
   },
   "_requiredBy": [
     "/@angular/cli"
   ],
-  "_resolved": "https://registry.npmjs.org/@schematics/angular/-/angular-7.1.4.tgz",
-  "_shasum": "cec4a5b793e54bd624018b6b5b4b917c467d22a5",
-  "_spec": "@schematics/angular@7.1.4",
-  "_where": "C:\\GrayAutomationApp\\automation-app\\node_modules\\@angular\\cli",
+  "_resolved": "https://registry.npmjs.org/@schematics/angular/-/angular-7.3.8.tgz",
+  "_shasum": "9cded496b79830164acbc2fc382b66e5b5f7cf02",
+  "_spec": "@schematics/angular@7.3.8",
+  "_where": "C:\\modules\\HomeAiAutomation\\automation_web_app\\automation-app\\node_modules\\@angular\\cli",
   "author": {
     "name": "Angular Authors"
   },
@@ -31,9 +31,9 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "@angular-devkit/core": "7.1.4",
-    "@angular-devkit/schematics": "7.1.4",
-    "typescript": "3.1.6"
+    "@angular-devkit/core": "7.3.8",
+    "@angular-devkit/schematics": "7.3.8",
+    "typescript": "3.2.4"
   },
   "deprecated": false,
   "description": "Schematics specific to Angular",
@@ -59,5 +59,15 @@
     "url": "git+https://github.com/angular/angular-cli.git"
   },
   "schematics": "./collection.json",
-  "version": "7.1.4"
+  "version": "7.3.8",
+  "warnings": [
+    {
+      "code": "ENOTSUP",
+      "required": {
+        "node": ">= 8.9.0",
+        "npm": ">= 5.5.1"
+      },
+      "pkgid": "@schematics/angular@7.3.8"
+    }
+  ]
 }
