@@ -1,28 +1,27 @@
 {
-  "_from": "strip-ansi@^4.0.0",
-  "_id": "strip-ansi@4.0.0",
+  "_from": "strip-ansi@^5.0.0",
+  "_id": "strip-ansi@5.2.0",
   "_inBundle": false,
-  "_integrity": "sha1-qEeQIusaw2iocTibY1JixQXuNo8=",
+  "_integrity": "sha512-DuRs1gKbBqsMKIZlrffwlug8MHkcnpjs5VPmL1PAh+mA30U0DTotfDZ0d2UUsXpPmPmMMJ6W773MaA3J+lbiWA==",
   "_location": "/inquirer/strip-ansi",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "strip-ansi@^4.0.0",
+    "raw": "strip-ansi@^5.0.0",
     "name": "strip-ansi",
     "escapedName": "strip-ansi",
-    "rawSpec": "^4.0.0",
+    "rawSpec": "^5.0.0",
     "saveSpec": null,
-    "fetchSpec": "^4.0.0"
+    "fetchSpec": "^5.0.0"
   },
   "_requiredBy": [
-    "/inquirer",
-    "/inquirer/string-width"
+    "/inquirer"
   ],
-  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-4.0.0.tgz",
-  "_shasum": "a8479022eb1ac368a871389b635262c505ee368f",
-  "_spec": "strip-ansi@^4.0.0",
-  "_where": "C:\\GrayAutomationApp\\automation-app\\node_modules\\inquirer",
+  "_resolved": "https://registry.npmjs.org/strip-ansi/-/strip-ansi-5.2.0.tgz",
+  "_shasum": "8c9a536feb6afc962bdfa5b104a5091c1ad9c0ae",
+  "_spec": "strip-ansi@^5.0.0",
+  "_where": "C:\\modules\\HomeAiAutomation\\automation_web_app\\automation-app\\node_modules\\inquirer",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
@@ -33,19 +32,21 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "ansi-regex": "^3.0.0"
+    "ansi-regex": "^4.1.0"
   },
   "deprecated": false,
-  "description": "Strip ANSI escape codes",
+  "description": "Strip ANSI escape codes from a string",
   "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+    "ava": "^1.3.1",
+    "tsd-check": "^0.5.0",
+    "xo": "^0.24.0"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=6"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
   "homepage": "https://github.com/chalk/strip-ansi#readme",
   "keywords": [
@@ -79,7 +80,7 @@
     "url": "git+https://github.com/chalk/strip-ansi.git"
   },
   "scripts": {
-    "test": "xo && ava"
+    "test": "xo && ava && tsd-check"
   },
-  "version": "4.0.0"
+  "version": "5.2.0"
 }
