@@ -1,7 +1,7 @@
 import { Component, OnInit	} from '@angular/core';
 import { ModelLive			} from '../model/ModelLive'
-import { FrameDataService   } from '../frame-data.service';
-import { ModelMain          } from '../model/ModelMain'
+import { FrameDataService	} from '../frame-data.service';
+import { CdkDragDrop } from '@angular/cdk/drag-drop';
 
 @Component({
   selector:     'app-live-feed',
@@ -12,10 +12,10 @@ import { ModelMain          } from '../model/ModelMain'
 export class LiveFeedComponent implements OnInit {
 	private model: ModelLive;
 
-	constructor(frameDataService: FrameDataService, modelMain: ModelMain)
-	{
-		this.model = new ModelLive(frameDataService, modelMain);
-		console.log('Constructing LiveFeedComponent');
+	constructor(frameDataService: FrameDataService)
+    {
+        console.log('Constructing LiveFeedComponent');
+		this.model = new ModelLive(frameDataService);
     }
       
     ngOnInit()
@@ -27,25 +27,18 @@ export class LiveFeedComponent implements OnInit {
         this.model.setPlaying();
     }
 
-    public changeRecordingIndex(index): void
-    {
-        this.model.setRecordingIndex(index);
-	}
-
 	public getCurrentFrameData	(): string	{ return					this.model.getCurrentFrameData();		}
 	public getTimeStamp			(): Date	{ return this.convertDate(	this.model.getCurrentFrameTimeStamp());	}
-	public getRecordings		(): any		{ return					this.model.getRecordings();				}
     public sourceActive         (): boolean { return                    this.model.sourceActive();              }
-    public getSliderPosition(): string
+
+	public getSliderPosition(): string
     {
         var percentage      = this.model.getSliderPercentage();
-        var pixelsFromLeft  = percentage * 648.0;                //TODO: can we get this value from the DOM?
-        var pixelString = pixelsFromLeft.toString() + "px";
-
-        console.log(pixelString);
+        var pixelsFromLeft  = percentage * 648.0;                                                                                   //TODO: can we get this value from the DOM?
+        var pixelString		= pixelsFromLeft.toString() + "px";
 
         return pixelString;
-    }
+	}
 
 	private convertDate(date): Date
 	{
